+++ NEWS.md
@@ -10,7 +10,8 @@
 * Additional CI methods (Agresti-Coull, Jeffreys) for `s_proportion`.
 * Added new CI methods `anderson-hauck` and `newcombe` to `s_proportion_diff`.
 * Additional p-value methods (Fisher's Exact, Chi-Squared Test with Schouten Correction) for `s_test_proportion_diff`.
-* The binary summary table function `t_binary_outcome` takes now lists (instead of character vectors) specified by the helper function `control_binary_comparison` as the arguments `strat_analysis` and `unstrat_analysis`. In addition, Odds Ratio estimates and CIs are now included by default and can be removed, similarly to the other subsections of the arm comparison analyses.
+* The binary summary table function `t_binary_outcome` takes now lists (instead of character vectors) specified by the helper function `control_binary_comparison` as the arguments `strat_analysis` and `unstrat_analysis`. In addition, Odds Ratio estimates and CIs are now included by default and can be removed, similarly to the other subsections of the arm comparison analyses. Also added argument `rsp_multinomial`.
+* Add new table function `t_el_multinomial_proportion`. 
   
 # tern 0.6.7